Redesigned the formatting of the user function chapter of the relax manual.
The fetch_docstrings.py now forces each user function to start in a new column.  This increases the
size of the manual, but makes the reading of the user function documentation much easier.  The user
function class and function icons (128x128 format) are now placed between the top bar and the 
subsection title and are left and right justified.  This prettification simply allows the user
functions to be more quickly identified.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@17347 b7916896-f9f9-0310-9fe5-b3996d8957d5