Python 3 fix for the relax_io.open_write_file() function.
This now matches the behaviour of open_read_file() in that there are three different behaviours for
opening bz2 and gz files for writing to for the different Python versions (one for Python 2, one for
Python 3.0 to 3.2, and one for Python 3.3+).  All byte streams have been eliminated as
open_write_file() is for creating text files.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@17718 b7916896-f9f9-0310-9fe5-b3996d8957d5