Added the 'empty' flag to the sequence.copy user function to allow all the spin contents to be copied.
The user function was only copying the basic empty molecule, residue and spin containers, in contrast
to the interatomic.copy user function which copies all of the container contents as well.  This new
flag is for backwards compatibility - it allows old scripts to operate as before while enabling the
new functionality.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18202 b7916896-f9f9-0310-9fe5-b3996d8957d5