Fix for the classic_style() method of the model-free specific Macro class.
This was identified by the new Mf.test_bug_20531_molmol_macro_write_relaxfault system test, but
would probably never be encountered by users.  It is not related to the bug this test is trying to
catch.  The problem is if spins have model-free parameters set up but there is no model-free model
initialised - then the Molmol and PyMOL macro creation would fail.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18514 b7916896-f9f9-0310-9fe5-b3996d8957d5