Fix for bug #20531 (https://gna.org/bugs/?20531).
This is the bug of the RelaxFault during the final stages of the auto_analysis.dauvergne_protocol
model-free analysis when the Molmol and PyMOL macros are being created.  The problem was caused by
model m9 (just the Rex parameter) when trying to determine what the order parameter for fast (< 200
ps) motions is.  The code was assuming that the model-free model would always contain an order
parameter.  It is very strange that this has only been rarely encountered!



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18531 b7916896-f9f9-0310-9fe5-b3996d8957d5