Fix for bug #20563 (https://gna.org/bugs/?20563), the end failure of the model-free auto-analysis.
This was reported by Stanislava Panova (https://gna.org/users/stacy).  The problem was related to
the Monte Carlo simulations.  The generic_fns.monte_carlo.create_data() function is now asking for
the errors associated with a certain data point, even if that data point does not exist.  The
solution was to check if the relaxation data ID string is present in spin.ri_data_err before
accessing the key-value pair (in the specific API _return_error_relax_data() common method).



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18726 b7916896-f9f9-0310-9fe5-b3996d8957d5