Fix for the bmrb.read user function for the recent molecule, residue and spin metadata improvements.
The generic_fns.bmrb.generate_sequence() function now calls generic_fns.mol_res_spin.metadata_clean()
to be sure that the metadata is correct.  The problem is the structure of the BMRB file with no spin
information in the entity record, hence the residues are created first and the spins much later in
generate_sequence().



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18751 b7916896-f9f9-0310-9fe5-b3996d8957d5