Fix for the system tests in the Mac OS X application binary.
The Mf.test_bug_20563_missing_ri_error system test fails in the Mac OS X application binary.  The
problem is that the py2app extension used to build the Mac application decides that empty
directories are not to be included in the app, so naturally the test fails when checking for these.
Now empty results files have been added to these directories to trick py2app to include them.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18788 b7916896-f9f9-0310-9fe5-b3996d8957d5