Fix for a strange and extremely rare typo bug in the model-free specific analysis code.
This was identified by Manish Chaubey <manish dott chaubey att tuebingen dott mpg dott de> in the
message at http://thread.gmane.org/gmane.science.nmr.relax.user/1422.

This only occurs if a relaxation data error of zero is encountered and is a bug in the RelaxError
message explaining the problem with the data.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18822 b7916896-f9f9-0310-9fe5-b3996d8957d5