Created the Structure.test_read_merge system test to test a new concept - merging of structures.
The idea is to add the merge argument to the structure.read_pdb user function to allow two different
structures in two PDB files to be merged.  This is useful if structures of individual domains have
have been solved separately and are located in two PDB files.  Then with the merge flag, you will
not need to use and external program or hand edit PDB files to join them.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18848 b7916896-f9f9-0310-9fe5-b3996d8957d5