Implemented the internal structural object _trim_helix() method.
This is used when the structure.delete user function is called to trim and remove the helix
metadata.  For this to work, the additional method _residue_data() was written to create a
dictionary with residue numbers as keys and the residue names as numbers.  This dictionary is used
by _trim_helix() to change the residue names in the helix metadata.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18882 b7916896-f9f9-0310-9fe5-b3996d8957d5