Fix for a nasty bug introduced at r18901.
The string decode() method was being called for the version.url() and version.revision() functions
to handle Python 3.  However this was causing strings in Python 2 to be converted to unicode, which
then caused the saving of XML state files to fail.  Now the version information is explicitly
converted to a string using the str() builtin function.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@18903 b7916896-f9f9-0310-9fe5-b3996d8957d5