Changed the 'Execute relax' button in all analyses in the GUI to 'Execute analysis'.
It makes no sense to execute relax as relax has been executing during the analysis initialisation
and the user setting up all the data for the analysis.  This is a remnant of ancient design of
Michael Bieri's GUI being a separate program to relax, which would execute relax with the click of
this button.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@19592 b7916896-f9f9-0310-9fe5-b3996d8957d5