Shifted the C array creation to the relax_fit C module header.
The params, values, sd, relax_times, and scaling_matrix C arrays are now declared and allocated in
the header file rather than using malloc() calls in the setup() function.  This is to attempt to
remove a memory leak.  The arrays are now of fixed length and reused for each setup() call.  These,
as well as the other variables declared in the header are no longer declared in the functions.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@19599 b7916896-f9f9-0310-9fe5-b3996d8957d5