The timing of individual tests in the relax test suite can now be printed out.
The new command line argument --time has been added which, when supplied with one of the test suite
arguments, will cause the time required to complete each individual test to be displayed.  Instead
of just printing the characters '.', 'F', and 'E' for each test, now these characters are postfixed
with the time in seconds, the name of the test and ending in a newline character.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@19703 b7916896-f9f9-0310-9fe5-b3996d8957d5