The spectrum.peak_intensities is now more robust when reading in a generic formatted file.
Firstly there is a check that the intensity column number has been supplied.  And then there is a
checks that all relevant data could be extracted from each row of the file.  This replaces traceback
errors with RelaxErrors explaining the problem if the user inputs bad data or forgets the intensity
column argument.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@19882 b7916896-f9f9-0310-9fe5-b3996d8957d5