Bug fix for the Sequence GUI input element.
This complete the removal of bug #20873 (https://gna.org/bugs/?20873).

The problem was that the gui_to_str() function was not failing to convert the string into a string
list, so the list was deemed as a single string.  Now the first character of '[' or '(' for lists or
tuples are now searched for instead of relying on the conversion to trigger an error.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@19947 b7916896-f9f9-0310-9fe5-b3996d8957d5