More MS Windows fixes, this time a nasty kludge, for the relax system tests.
This is strangely needed for the relax_disp branch and not the trunk for a 64-bit MS Windows 7 test
system.  The reason why this WindowsError is triggered by the base tearDown() method in the
relax_disp branch and not trunk is a total mystery.  Actually why Windows refuses to complete the
file close() operations of the results.write and state.save user functions before calling the
tearDown() method is the greater mystery.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@20211 b7916896-f9f9-0310-9fe5-b3996d8957d5