The relax language definition is now auto-generated by the fetch_docstrings.py script.
This is for use in the relax user manual using the listings package.  The fetch_docstrings.py
script now creates the docs/latex/script_definition.tex file.  This is used by the relax.tex file
via an \include{} statement.  This setup allows all of the relax user functions to be dynamically
set as keywords for the relax language definition.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@20235 b7916896-f9f9-0310-9fe5-b3996d8957d5