Creation of the optimisation chapter of the relax user manual.
The main text of this chapter originates as part of the model-free chapter.  As this most of this
text was not model-free specific, it has been spun out as its own chapter.  Text has also been taken
from the "Optimisation of relaxation data – values, gradients, and Hessians" chapter.

The indexing for the optimisation topics has also been improved.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@20526 b7916896-f9f9-0310-9fe5-b3996d8957d5