Large change to the free file format GUI element for the user functions.
The GUI element used in the user function wizard windows has been modified to have both a 'default'
form, which is the previous design, and a 'mini' form which is now used for the user functions.
This mini form only uses 1 row, rather than the default of 6 or 8.  It is a read only text element
with a button that launches the free file format window.  The amount of space saved is huge.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@20582 b7916896-f9f9-0310-9fe5-b3996d8957d5