Added support for the Mac OS X Modelfree4 binary results to the Palmer.* system tests.
The Mac OS X Modelfree 4.20 binary produces different results than the Linux binaries, mainly due to
a compilation problem.  In the Linux binaries, the results are written out to 4 decimal places.  In
the Mac binaries, the results are instead written out to 4 significant figures.  Therefore the
number of decimal places are much less than the Linux results.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@20594 b7916896-f9f9-0310-9fe5-b3996d8957d5