Added multi-processor support for Monte Carlo simulations.
This simply involves accessing the multi-processor box singleton and running the
processor.run_queue() method within the pipe_control.minimise.minimise() function.  This currently
does nothing as the processor queue is always empty.  But if the code in the specific_analyses
package is modified to add slave commands to the processor but not execute the run_queue() method,
then the Monte Carlo simulations will be automatically parallelised.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@21012 b7916896-f9f9-0310-9fe5-b3996d8957d5