Fix for the Monte Carlo simulation parallisation.
The processor.run_queue() method needs to be executed for all the optimisation user function of
relax (at the end of both the grid_search() and the calc() functions).  Otherwise when removing the
processor.run_queue() call from the specific_analyses code results in a failure of the grid search
or back-calculation to be run until the minimise() user function call, which will be too late.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@21015 b7916896-f9f9-0310-9fe5-b3996d8957d5