Fix for bug #21187 - the corrupted PDB issue with protons atom numbers of zero.
The bug at https://gna.org/bugs/?21187 was reported by Martin Ballaschk (https://gna.org/users/mab).

The fix was to allow for spin containers in the relax data store to have the same atom number, as
long as the atom names are different.  Note - this may cause problems in the future so a better
solution might have to be found.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@21051 b7916896-f9f9-0310-9fe5-b3996d8957d5