Started to add better pseudo-atom support for the PCS.
The new pipe_control.pcs.setup_pseudoatom_pcs() function has been added to deselect the spins which
are members of a pseudo-atom.  The return_pcs_data() function in the same module now calls this
function and builds a list of pseudo-atom flags for use in the target function (though it is still
unused).



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@21450 b7916896-f9f9-0310-9fe5-b3996d8957d5