Modified the 'NS R1rho 2-site' dispersion model to match the 'NS R1rho 3-site' models.
The 6D evolution matrix indices have been rearranged to match the 9D matrix indices.  The tilt angle
for the initial magnetisation is no longer that for the average offset but that of state A, as was
changed for the 'NS R1rho 3-site' models earlier (r21911).  The system test was therefore updated
for the slightly different behaviour.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@21916 b7916896-f9f9-0310-9fe5-b3996d8957d5