Fix for bug #21522 (https://gna.org/bugs/?21522) and bug #21520 (https://gna.org/bugs/?21520).
These bugs are the structure.write_pdb user function creating an incorrect MASTER record and the
failure of the structure.write_pdb user function when creating the MASTER record due to too many
ATOM and HETATM records being present.

The counts for the ATOM, HETATM, and TER records are now only for a single model, rather than being
the sum for all models together.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22034 b7916896-f9f9-0310-9fe5-b3996d8957d5