Yet more optimisation for handling large quantities of models in the internal structural model.
Now when adding new models to the object, the model_indices and model_list objects are no longer
created.  This saves much time as the large model_list is now not sorted.  A number of structural
object methods have been updated to handle the change by switching to the model_loop() method for
looping over the models, rather than using the model_indices and model_list objects.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22053 b7916896-f9f9-0310-9fe5-b3996d8957d5