Fix for bug #21562, the failure of the NOE analysis when spectra are replicated.
This bug (https://gna.org/bugs/?21562) was reported by Dhanas Muthu (https://gna.org/users/dhanas).

The problem was that the NOE overfit_deselect() method was deselecting all spins which do not have
exactly 2 intensity values.  This is incompatible with replicated spectra as the number will be
greater than two.  The check has been modified to deselect spins only when the number of intensity
values are zero or one (i.e. < 2 rather than != 2).



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22088 b7916896-f9f9-0310-9fe5-b3996d8957d5