Better support for replicated spectra in the NOE analysis.
The saturated and reference peak intensity and error are now properly averaged.  Previously
averaging was not used as the number of replicates N are cancelled in the ratios used for the NOE
and error calculation.  However this fails when the number of replicates for the saturated spectrum
does not match the number of replicates for the reference spectrum.  Now any data combination is
possible.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22090 b7916896-f9f9-0310-9fe5-b3996d8957d5