Modified the help text to explain that -p will launch relax in prompt mode after running any optionally supplied scripts.
Fix for sr #3117, (https://gna.org/support/?3117) - Functionality to inspect interactively after running script - The equivalence to python -i

This is to allow the -p --prompt option to be given without a script. relax should support this so that a user doesn't get too confused when
trying to start in prompt mode with the --prompt flag and support the --prompt argument without a script being supplied.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22186 b7916896-f9f9-0310-9fe5-b3996d8957d5