Complete reformatting of the base LaTeX files.
The paragraph structure has been changed so that each sentence now starts on a new line.  This is
for better tracking of changes (via 'svn diff' for example), for better searchability of certain
text elements using command line tools such as 'grep', and for easier easier use of the 'sed' tool.
The change tracking is most important as it allows for finer granularity - a small change will now
only be shown as a change in one sentence rather than the whole paragraph, allowing the change to be
identified more easily.  It also allows for easier commit maintenance.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22210 b7916896-f9f9-0310-9fe5-b3996d8957d5