Added system test for bug 21344.
Regarding bug #21344, (https://gna.org/bugs/index.php?21344) - Handling of in sparse acquired R1rho dataset with missing combinations of time and spin-lock field strengths.

This test will fail with:
No intensity data could be found corresponding to the spectrometer frequency of 799.7773991 MHz, dispersion point of 431.0 and relaxation time of 0.14 s.
Data for a dispersion point of 431.0 and time 0.14 does not exist, and so some of the looping in collecting data for calculation must be wrong.
This behaviour and probably its solution is related to bug 21665, "Running a CPMG analysis with two fields at two delay times" (https://gna.org/bugs/index.php?21665).

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22321 b7916896-f9f9-0310-9fe5-b3996d8957d5