Fix for bug #21754 (https://gna.org/bugs/?21754).
This is the failure of the grace.view user function on MS Windows in opening a file in qtgrace when
the path includes spaces, as reported by Mengjun Xue (mengjun dott xue att mailbox dott tu-berlin
dott de).

The fix was to run the program with the file name in double quotes.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22405 b7916896-f9f9-0310-9fe5-b3996d8957d5