Fix for bug #21814 (https://gna.org/bugs/?21814).
This is the PDB reading failure when the PDB records are not padded to 80 spaces.  The fix is
simple, all PDB records are pre-validated.  This includes removing all newline characters and
padding each PDB record to 80 spaces when needed.  This will however add an overhead cost -- the
internal PDB reader will now be slower.  However corrupted PDB files, produced by MODELLER for
example, not padded to 80 spaces will now be better supported.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22509 b7916896-f9f9-0310-9fe5-b3996d8957d5