Added the new exp_mc_sim_num argument to the relaxation dispersion auto-analysis.
This is in preparation for fixing bug #21869 (https://gna.org/bugs/?21869).

This argument allows for a different number of Monte Carlo simulations for the 'R2eff' model when
exponential curves are fit.  It will mainly be useful in the test suite to improve the accuracy of
the R2eff errors, while still running a low number of simulations for the other models to allow
optimisation to be quick.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22565 b7916896-f9f9-0310-9fe5-b3996d8957d5