Modifications to the Relax_disp.test_m61_exp_data_to_m61 system test.
This is to fix bug #21869 (https://gna.org/bugs/?21869), the failure of this system test.

The number of Monte Carlo simulation for the 'R2eff' model has been increased from 3 to 25 using the
new exp_mc_sim_num argument to the dispersion auto-analysis.  To keep the test fast, only a single
spin is optimised.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22566 b7916896-f9f9-0310-9fe5-b3996d8957d5