Standardised all of the parameter list objects for the specific analyses.
All of the specific_analyses packages now contain a parameter_object module which defines all of the
parameters.  The ordering of these is now consistent between the analyses, the result of which will
be more consistent ordering of element in the relax XML state files.

The new Param_list.add_csa() base class method has been added to standardise the CSA parameter.  And
the capabilities of the add_model_info() method has been expanded.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22670 b7916896-f9f9-0310-9fe5-b3996d8957d5