The frame order pivot points are now stored differently in the current data pipe.
Instead of being stored as a list or array of numbers in cdp.pivot, the point is now stored as
cdp.pivot_x, cdp.pivot_y and cdp.pivot_z.  The second pivot cdp.pivot2 is now stored as
cdp.pivot_x_2, cdp.pivot_y_2 and cdp.pivot_z_2.

This is to simplify the automated handling of optimisation parameters.  Rather than having to
convert the pivot_x, pivot_y, and pivot_z parameters to and from a list, the same mechanisms can now
be used for all of the optimised frame order parameters.  This will be used to hugely simplify many
of the functions in the specific_analyses.frame_order.parameters module and eliminate a large source
of bugs.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22672 b7916896-f9f9-0310-9fe5-b3996d8957d5