Added the ability to automatically create the parameter tables for the user function documentation.
These are the tables used in many of the user function docstrings.  This has been added to the
parameter list base class.  The section title is pre-specified by the new _set_uf_title() method,
and the table caption and LaTeX label by the _set_uf_table() method.  The documentation is generated
by calling the uf_doc() method.  This uses the new type_string() method to add a compact parameter
type string representation to the table.

The aim is to eliminate all of the hard-coded tables in the specific analyses which are always very
quickly out of date.  By automatically creating the tables, this simplifies the codebase and
simplifies the addition of new analysis types.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22718 b7916896-f9f9-0310-9fe5-b3996d8957d5