Shifted the data_store.relax_xml module into the relax library as lib.xml.
This module contains a couple of functions which are used for converting Python objects into an XML
representation and back again.  These are used not only by the relax data store, but also a number
of the structural objects in the relax library (which are themselves placed in the relax data
store).  This makes the relax library more independent from the rest of relax.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22789 b7916896-f9f9-0310-9fe5-b3996d8957d5