Created a new test category for the relax test suite - the software verification tests.
This is part of the full test suite or can be run by itself using the new --verification-tests
command line option.  Such tests are best described by
https://en.wikipedia.org/wiki/Verification_and_validation_%28software%29.  These tests will be used
to make sure that the design aims of the relax source code are satisfied.  For example that the
relax library is independent from the rest of relax.  Or that the package __all__ lists actually
contain all modules and sub-packages (these tests are currently part of the unit tests).



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22793 b7916896-f9f9-0310-9fe5-b3996d8957d5