Expanded the relaxation dispersion auto-analysis to allow the grid search to be turned off.
By setting the grid_inc argument to None, the grid search will be turned off.  As a replacement, the
value.set user function is used for all model parameters to set them to their default values prior
to minimisation.  This design is for speed as optimisation from the defaults is often - though not
always - good enough.  It can be used, for example, in the test suite to make the system tests much
faster.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22848 b7916896-f9f9-0310-9fe5-b3996d8957d5