Made a generic script to generate R2eff data for a CPMG model based on spin parameters, and fit the data.
Still needs to implement some noise method.

1) The idea is to generate R2eff data with a numerical model with some extreme parameters.
2) Then add noise to the R2eff data.
3) Then fit with a analytical model.

Evaluate the performance on the analytical model.
This follows the idea of the paper:

http://dx.doi.org/10.1016/j.jmr.2014.02.023
"An exact solution for R2,eff in CPMG experiments in the case of two site chemical exchange"
Andrew J. Baldwin, Journal of Magnetic Resonance, 2014

The script can be extended to also include global fitting, to test this out.

The script is also ideal, when trying to implement a new model, since test-data is ready at hand.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@22909 b7916896-f9f9-0310-9fe5-b3996d8957d5