Latex2html fixes for the HTML version of the relax manual.
This is for the documentation at http://www.nmr-relax.com/manual/index.html.

Latex2html has problems determining if the contents of environments should added to the sub or
superscript.  For example $1^\textrm{st}$ is not recognised and must be changed to $1^{\textrm{st}}$
for latex2html to function correctly.  Therefore these problems have been fixed throughout the
manual.  The number of errors printed out by latex2html is now significantly less.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23129 b7916896-f9f9-0310-9fe5-b3996d8957d5