Increased the precision of many of the Frame_order.test_rigid_data_to_*_model system tests.
This is to fix a test which was failing due to the recent re-parameterisation of the rotor frame
order model to eliminate one parameter.  The precision of the numeric Sobol' sequence integration
has been increased by shifting the fixed parameter values even closer to zero.  As a consequence,
the chi-squared value of five of these tests is now lower.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23142 b7916896-f9f9-0310-9fe5-b3996d8957d5