Big bug fix for the relax installation path determination.
This is to fix bug #22037 (http://gna.org/bugs/?22037), the failure to load graphics in the GUI due
to the relax installation path not being set up correctly.

The problem is that the status module was looking for the compat.py file to determine where the base
directory is, but this file has been moved into the lib/ package.  Now the dep_check.py file is
being searched for.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23157 b7916896-f9f9-0310-9fe5-b3996d8957d5