Improvements for the 2D Grace plots created by the rdc.corr_plot user function.
The autoscaling is now turned off, as the data set representing the diagonal (with points
[-100, -100] and [100, 100]) causes the world view to be set to be between -100 to 100 or -200 to
200.

The world view is set to be between -50 and 50 Hz, so that all RDCs should be visible.  The ticks in
the plot have also been set so that the minor ticks are at every Hz increment.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23167 b7916896-f9f9-0310-9fe5-b3996d8957d5