Fix for bug #22041, the PDB atom serial number error from the structure.write_pdb user function.
This was reported at https://gna.org/bugs/?22041.

The problem is that the structure.write_pdb user function preserves the atom numbering from the
original structure and uses that for the atom serial number.  However the atom serial number must be
replaced with sequential values to produce a valid PDB file.  This is fatal for any CONECT records.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23188 b7916896-f9f9-0310-9fe5-b3996d8957d5