Modified the 2nd rotor model of CaM frame order optimisation script.
The frame_order.pdb_model user function is now used to create a PDB representation of the rotor
motions for the real, expected parameters and for the optimisation results when the pivot point is
fixed.

In addition, the pymol.cone_pdb user function has been replaced by the pymol.frame_order user
function.  All new files have been added to the repository.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23191 b7916896-f9f9-0310-9fe5-b3996d8957d5