Completed the frame_order.pdb_model user function backend for the frame order PDB representation.
Most of this backend, including the axes and cone representations, had been broken for quite a while
and were being skipped with an early return statement.  This has now been made functional and a few
fixes have been made.

For the 'rotor' and 'free rotor' model, the neg_cone argument is now ignored so that only one model
is produced in the final frame order PDB representation file.  For all other models, the rotor
representation is no longer centred to the point on axis closest to the centre of mass, as the pivot
is unambiguously defined.  The rotor representation has also been made larger in these models so
that it is outside of the cone, and the propeller blades are now staggered.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23476 b7916896-f9f9-0310-9fe5-b3996d8957d5