Created the lib.compat.norm() compatibility function for numpy.linalg.norm().
For numpy 1.8 and higher, the numpy.linalg.norm() function has introduced the 'axis' argument.  This
is an incredibly fast way of determining the norm of an array of vectors.  This is used by the frame
order analysis.

However for older numpy versions, this causes the frame order analysis, and many corresponding
system and GUI tests to fail.  Therefore this new lib.compat.norm() function has been designed to
default to numpy.linalg.norm() if the axis argument is supported, or to switch to the much slower
numpy.apply_along_axis(numpy.linalg.norm, axis, x) call which is supported by older numpy.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23604 b7916896-f9f9-0310-9fe5-b3996d8957d5