The built in Python range() function is no longer being replaced by xrange().
Replacing builtin.range() with builtin.xragne() on Python 2 was causing problems with Python
site-packages which were not Python 3 compliant.  This includes old numpy versions.

The original overwriting of range() with xrange() was for both speed and memory conservation.
However profiling the system tests, the time for all tests did not change significantly.

This change may cause problems in certain places in relax on memory constrained computer systems,
so it may need to be reverted in the future.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@23606 b7916896-f9f9-0310-9fe5-b3996d8957d5