Numpy version 1.6 or higher is now required to be able to run relax.
This follows from the series of messages:

- http://www.mail-archive.com/relax-devel@gna.org/msg06288.html,
- http://www.mail-archive.com/relax-devel@gna.org/msg06289.html,
- http://www.mail-archive.com/relax-devel@gna.org/msg06327.html,
- http://www.mail-archive.com/relax-devel@gna.org/msg06335.html.

If too many users complain, maybe this change can be reverted later.  This minimal numpy version is
needed for many of the speed ups going in the relaxation dispersion and frame order analyses.  It is
required for the numpy ufunc out arguments and for the numpy.eigsum() function.  These will likely
be used in other analyses in the future for improving the speed of relax, so it might affect users
of other analyses later on.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@24354 b7916896-f9f9-0310-9fe5-b3996d8957d5