Fix for bug #22257, the freezing of the GUI after using the free file format window on Mac OS X.
This is reported at https://gna.org/bugs/?22257.

This is a recurring problem in Mac OS X as it cannot be tested in the relax test suite.  The problem
is with wxPython.  The modal dialogs, such as the free file format window, cannot be destroyed on
Mac OS X using wx.Dialog.Destroy() - this kills wxPython and hence kills relax.  The problem does
not exist on any other operating system.  To fix this, all wx.Dialog.Destroy() calls have been
replaced with wx.Dialog.Close().



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@24387 b7916896-f9f9-0310-9fe5-b3996d8957d5