Updated the Mac OS X framework setting up instruction document.
New sections have been added for the nose and matplotlib Python packages, as nose is needed for the
numpy and scipy testing frameworks and matplotlib might be a useful optional dependency in the
future.

The mpy4py section has been updated to avoid the non-framework fink version of mpicc which cannot
produce universal binaries.  A few other parts also have small edits.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@24425 b7916896-f9f9-0310-9fe5-b3996d8957d5