Made the GUI selection of models for relaxation dispersion more simple.
After the implementation of a function which will translate the models, the
'No Rex' model will be converted to the 'No Rex' model for R1rho off-resonance.

Also the corresponding 'R1 fit' model will be chosen istead, if R1 data has not been loaded.

This makes the model selection easier in the GUI interface.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25100 b7916896-f9f9-0310-9fe5-b3996d8957d5