Modified intermediate systemtest: Relax_disp.test_bug_9999_slow_r1rho_r2eff_error_with_mc
to see if the initial Grid Search for 'i0' and 'R2eff' estimation can be skipped.

This is done by converting the exponential curve, to a linear curve, and calculate
the best parameters by a line of best fit by least squares.

This seems like a promising method as an initial estimator of 'i0' and 'r2eff'.

For 500 to 2000 Monte-Carlo simulations, this could have an influence on the timings,
as all grid searchs could be skipped.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25189 b7916896-f9f9-0310-9fe5-b3996d8957d5