Fix for bug #22502 (https://gna.org/bugs/?22502).
This is the problem whereby the geometric prolate diffusion representation does not align with axis
in PDB, as reported by Martin Ballaschk (https://gna.org/users/mab).

This problem was not the main prolate tensor axis, but that the geometric object needed to be
rotated 90 degrees about the z-axis to bring the object and axis into the same frame.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25206 b7916896-f9f9-0310-9fe5-b3996d8957d5