Moved target function for curve fitting with scipy into specific_analyses.relax_disp.estimate_r2eff.
This will later include the backend specific_analyses.relax_disp.optimisation.estimate_r2eff()
function and the code in the target_functions package.

The code in target_functions.relax_disp_curve_fit is a lot more than just a target function, so it doesn't really belong in this package.

This is also to isolate this experimental feature.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25243 b7916896-f9f9-0310-9fe5-b3996d8957d5