Standardisation of the array dimensionality in the target function C code.
The new target_functions/dimensions.h header file defines MAX_PARAMS and MAX_DATA which is then
included in the header files of all the other C files.  All array declarations now explicitly
specify the length of each dimension.

The values of MAX_PARAMS and MAX_DATA have increased from 3 and 50 to 20 and 5000.  This is to allow
for models with more parameters and to allow a much larger number of input data points to be
supported before memory corruptions happen.  The data structures now take up more memory, but as the
functions do not loop up to maximum but only the number of parameters and points specified, this
will not make the code slower.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25305 b7916896-f9f9-0310-9fe5-b3996d8957d5