Implemented the specific_analyses.relax_fit.optimisation.d2func_wrapper() function.
This converts the numpy parameter array into a Python list, calls the
target_functions.relax_fit.d2func() function with this list, converts the Hessian output list of
lists into a numpy rank-2 array, and returns it.  This will allow Newton optimisation to be used for
the relaxation curve-fitting analysis.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25315 b7916896-f9f9-0310-9fe5-b3996d8957d5