Added keyword "exp_mc_sim_num", to the auto analyses in the GUI.
This sets the number of Monte Carlo simulations for R2eff error estimation in exponential curve fitting.

When setting to -1, the errors are estimated from the Covariance Matrix.

These errors are highly likely to be wrong, but can be used in Rapid testing of data and plotting.

task #7822(https://gna.org/task/index.php?7822): Implement user function to estimate R2eff and associated errors for exponential curve fitting.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25361 b7916896-f9f9-0310-9fe5-b3996d8957d5