Added isotope information to the lib.periodic_table module including mass number and atomic mass.
A new Isotope data container has been added to store this information.  The Periodic_table._add()
method now returns the initialised Element container.  This container has the _add_isotope() method
which is used to initialise Isotope data containers with the mass number and atomic mass and append
it to the list.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25380 b7916896-f9f9-0310-9fe5-b3996d8957d5