First attempt at properly implementing the target_functions.relax_fit.jacobian() function.
This is now the Jacobian of the chi-squared function.  A new jacobian_matrix data structure has been
created for holding the matrix data prior to converting it into a Python list of lists.  The
equation used was simply the chi-squared gradient whereby the sum over i has been dropped and the i
elements are stored in the second dimension of matrix.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25413 b7916896-f9f9-0310-9fe5-b3996d8957d5