Correction for catastrophic implementation of Monte-Carlo simulations.
And wrong implemetented "else if" statement, would add the intensity for the simulated intensity together with
the original intensity.

This means that all intensity values send to minimisation would be twice as high than usually.
(If spectra was not replicated.)

This was discovered for Monte-Carlo simulations of R2eff errors in exponential fit.

The function is restricted to the analysis of errors for exponential fit in Relax Dispersion.
Such data are normally restricted to R1rho analysis.

This will affect all analysis of R1rho data performed until now.
By pure luck, it seems that the effect of this would be that R2eff errors are half the value they should be.
A further investigation shows, that for the selected data set, this had a minimum on influence on the fitted parameters,
because the chi2 value would be scaled up by a factor 4.

bug #22554(https://gna.org/bugs/index.php?22554): The distribution of intensity with errors in Monte-Carlo simulations are markedly more narrow than expected.
task #7822(https://gna.org/task/index.php?7822): Implement user function to estimate R2eff and associated errors for exponential curve fitting.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25491 b7916896-f9f9-0310-9fe5-b3996d8957d5