Merged the profiling_cr72.py and profiling_dpl94.py scripts into base.py.
These dispersion model profiling scripts have been merged into one base script, and the code in
profiling_cr72.py and profiling_dpl94.py minimised to imports, two class initialisation lines, and
the execution of the main() function.  This removes a huge amount of code duplication and allows the
profiling scripts to be more easily modified, for example for compatibility with new relax versions.
All of the other profiling scripts will be updated to use this base script.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25617 b7916896-f9f9-0310-9fe5-b3996d8957d5