Added all user function ranamings since relax 1.3.1 to the translation table.
These were taken directly from the docs/user_function_changes.txt document.  Earlier relax versions
are far too different, so this will be the earliest relax version for this translation table.  The
relax 1.2 and earlier (and 1.3.0) versions used the run argument throughout and the scripting was so
different, that telling the user how to upgrade to new user functions is pointless.  And the release
date of relax 1.2.15, the last of these old designs was in November 2008.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25701 b7916896-f9f9-0310-9fe5-b3996d8957d5