Imported the Numdifftools 0.6.0 package into the relax source tree.
This package is extremely useful for testing the implementation of gradients, Hessians, and
Jacobians for all relax target functions.  The numerical values from numdifftools can be compared to
the directly calculated values.  And for analysis types where the partial derivatives with respect
to each model parameter are too complicated to calculated, or the derivatives are very complicated
and hence slow, numdifftools can be used to provide a numerical estimate for direct use in the
optimisation.

The Numdifftools package is from https://pypi.python.org/pypi/Numdifftools and
https://code.google.com/p/numdifftools/.  The current version 0.6.0 has been placed into
extern/numdifftools.  This is only the numdifftools package within the official distribution files
and the Python package setup.py file and associated files and directories have not been included.
The package uses the New BSD licence (the revised licence with no advertising clause) which is
compatible with the GPL v3 licence.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25898 b7916896-f9f9-0310-9fe5-b3996d8957d5