Shifted the specific_analyses.relax_disp.variables module into lib.dispersion.
This is both to minimise circular dependencies, as previously the specific_analyses.relax_disp
modules import from target_functions.relax_disp and vice-versa, and to allow the relax library
functions to have access to these variables.  This follows from a similar change to the frame order
analysis in the frame_order_cleanup branch.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25965 b7916896-f9f9-0310-9fe5-b3996d8957d5