Converted all structural object methods to use the selection object rather that atom ID strings.
This should have a significant impact on the speed of certain operations within relax.  The most
obvious effect will be a huge speed up of the interatom.define user function.  There should be speed
ups with a number of other user functions relating to structural information.  All parts of relax
have been updated for the change.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@25979 b7916896-f9f9-0310-9fe5-b3996d8957d5