Changed the design of the Check object in lib.checks.
The design of the checking function to call has been modified - it should now return either None if
the check passes or an instantiated RelaxError object if not.  This is then used to determine if the
__call__() method should return True (when None is received).  Otherwise if escalate is set to 1,
the text from the RelaxError object is sent into a RelaxWarning and False is returned.  And if
escalate is set to 2, then the error object is simply raised.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26055 b7916896-f9f9-0310-9fe5-b3996d8957d5