Fix for the experimental information data pipe object when converting to XML state and results files.
This is a partial fix for bug #22704 (https://gna.org/bugs/?22704), the corrupted relax state files
after setting the relax references via the bmrb.software, bmrb.display, or bmrb.write user
functions.

The names and descriptions for the software, citation and script list objects were incorrectly set.
These have been fixed so that the name of the data structure and the real description is present in
the XML state or results file instead of <relax_list desc='relax list container'>.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26067 b7916896-f9f9-0310-9fe5-b3996d8957d5