Modified the printouts for the unit tests when running with the --time command line option.
The test name is now being processed.  The leading 'test_suite.unit_tests.' text is now stripped
out.  And the remaining text is split into the module name and the test name.  This is to allow the
unit test module name to be more easily identifiable, so it can then be used as a command line
option to allow only a subset of tests to be performed.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26191 b7916896-f9f9-0310-9fe5-b3996d8957d5