Critical fix, for k_AB not belonging to list of global parameters.
k_AB was only changed to the spin of interest, but not for the rest of the cluster.
When the parameter vector is assembled,  "assemble_param_vector(spins=spins)" it takes the global parameter from spin 0.

Bug #22754 (https://gna.org/bugs/index.php?22754): The minimise.calculate() does not calculate chi2 value for clustered residues.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26197 b7916896-f9f9-0310-9fe5-b3996d8957d5