Big bug fix for the text size formatting of the HTML manual.
The previous fix for the user function chapter of the HTML manual
(http://www.nmr-relax.com/manual/Alphabetical_listing_user_functions.html) did not fix the problem.

The issue was with the {exampleenv} defined using a \newenvironment command in the preamble.  The
command \footnotesize was bing used in the start, but nothing was changing the font size at the end.
In LaTeX, the ending of the environment appears to reset the font size, whereas in latex2html it
does not.  Therefore all text after this environment is prepended by <SMALL CLASS="FOOTNOTESIZE"> in
the HTML manual and this keeps adding to the text after each new exampleenv environment.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26234 b7916896-f9f9-0310-9fe5-b3996d8957d5