Created systemtest Relax_disp.test_dx_map_clustered_create_par_file, which must show that relax is not able to find the local minimum under clustered conditions.
When creating the map, the map contain chi2 values, which are lower than the clustered fitted values.

This should not be the case.

I am running a larger map with larger bounds and more increements, which should show that there exist a minimum in the minimisation space with a lower chi2 value.

Bug #22754 (https://gna.org/bugs/index.php?22754): The minimise.calculate() does not calculate chi2 value for clustered residues.
Task #7860(https://gna.org/task/index.php?7860): When dx_map is issued, create a parameter file which maps parameters to chi2 value.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26244 b7916896-f9f9-0310-9fe5-b3996d8957d5