Added the write out of a matplotlib command file, to plot surfaces of a dx map.
It uses the minimum chi2 value in the map space, to define surface defitions.

It creates a X,Y; X,Z; Y,Z map, where the values in the missing dimension has been cut at the minimum chi2 value.
For each map, it creates a projected 3d map of the parameters and the chi2 value, and a heat map for the contours.

It also scatters the minimum chi2 value, the 4 smallest chi2 values, and maps any points in the point file, to a scatter point.

Mapping the points from file to map points, is done by finding the shortest Euclidean distance in the space from the points to any map points.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26258 b7916896-f9f9-0310-9fe5-b3996d8957d5