Added more checks to the determine_rnd() of the dauvergne_protocol model-free auto-analysis.
This is to try to catch bizarre situations such as bug #22730 (https://gna.org/bugs/?22730),
model-free auto-analysis - relax stops and quits at the polate step.

The following additional fatal conditions are now checked for:  A file with the same name as the
base model directory already exists;  The base model directory is not readable;  The base model
directory is not writable.  The last two could be caused by file system corruptions.  In addition,
the presence of the base model directory is checked for using os.path.isdir() rather than catching
errors coming out of the os.listdir() function.  These changes should make the analysis more robust
in the presence of 'strangeness'.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26283 b7916896-f9f9-0310-9fe5-b3996d8957d5