Changed the logic for how the new analysis wizard in the GUI is destroyed.
This relates to bug #22818 (https://gna.org/bugs/?22818), the GUI test suite failures in MS Windows
- PyAssertionError: C++ assertion "Assert failure".

The Destroy() method has been added to the Analysis_wizard class to properly close all elements of
the wizard.  This is now called from the menu_new() method of the Analysis_controller class, which
is the target of the menu item and toolbar button.

To allow the test suite to use this, the menu_new() method now accepts the destroy boolean argument.
The test suite can set this to False and then access the GUI elements after calling the method
(however the Destroy() method must be called by the test suite).



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26305 b7916896-f9f9-0310-9fe5-b3996d8957d5