Resign of how the new analysis wizard is handled in the GUI tests.
This relates to bug #22818 (https://gna.org/bugs/?22818), the GUI test suite failures in MS Windows
- PyAssertionError: C++ assertion "Assert failure".

The GUI test base class method new_analysis_wizard() has been created to simplify the process.  When
a new analysis is desired, this method should be called.  It will return the analysis page GUI
element for use in the test.  The method standardises the execution of the new analysis wizard and
sets up the analysis in the GUI.  It also properly destroys the wizard to avoid the memory leaking
issues such as bug #22818.

All GUI tests have been converted to use new_analysis_wizard().  This allows the GUI tests to pass
on MS Windows.  However there are still significant sources of memory leaks (the USER Objects count)
visible in the Windows Task Manager.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26306 b7916896-f9f9-0310-9fe5-b3996d8957d5