Implemented the multiple molecule merging functionality of the structure.load_spins user function.
The argument has been added to the user function frontend and a description added for this new
functionality.  In the backend, the pipe_control.structure.main.load_spins() function will now call
the load_spins_multi_mol() function if from_mols is supplied.  This alternative function is required
to handle missing atoms and differential atom numbering.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26327 b7916896-f9f9-0310-9fe5-b3996d8957d5