Fix for the structure.load_spins user function when the from_mols argument is used.
The load_spins_multi_mol() function of the pipe_control.structure.main module was incorrectly
handling the atomic position returned by the internal structural object atom_loop() method.  This
position is a list of lists when multiple models are present.  But when only a single model is
present, it returns a simple list.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26368 b7916896-f9f9-0310-9fe5-b3996d8957d5