Bug fix for the structure.load_spins user function new from_mols argument.
This was incorrectly using the pipe_control.pipes.pipe_names() function to obtain its default values
in the GUI (although this is not currently uesd).  The result was a non-fatal error message on Mac
OS X systems of "Python[1065:1d03] *** __NSAutoreleaseNoPool(): Object 0x3a3944c of class NSCFString
autoreleased with no pool in place - just leaking".



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26557 b7916896-f9f9-0310-9fe5-b3996d8957d5