Bug fix for the lib.structure.angles.angles_*() functions for odd increments.
This affects the PDB representations of the diffusion tensor and frame order when the number of
increments in the respective user functions is set to an odd number.  It really only affects the
frame_order.pdb_model user functions, as the number of increments cannot be set in any of the other
user functions (structure.create_diff_tensor_pdb, structure.create_rotor_pdb,
structure.create_vector_dist, n_state_model.cone_pdb).



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@26681 b7916896-f9f9-0310-9fe5-b3996d8957d5