Improvements for all of the tables of the relaxation dispersion chapter of the manual.
The captions are now the full width (or height for rotated tables) of the page in the PDF version of
the manual.  The \latex{} command from the latex2html package has been used to improve the HTML
versions of the tables by deactivating the landscape environment, the cmidrule command, and the
caption width commands.  This results in properly HTML formatted tables, rather than creating a PNG
image for the whole table.  These should significantly improve the tables in the webpages
http://www.nmr-relax.com/manual/Comparison_of_dispersion_analysis_software.html,
http://www.nmr-relax.com/manual/The_relaxation_dispersion_auto_analysis.html, and
http://www.nmr-relax.com/manual/Dispersion_model_summary.html.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27044 b7916896-f9f9-0310-9fe5-b3996d8957d5