Change to the Structure.test_align system test.
The molecules argument for the structure.align user function has been changed to match the models
argument, in that it now needs to be a list of lists with the first dimension matching the pipes
argument.  This change is to help with the implementation of the new structure.align functionality.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27056 b7916896-f9f9-0310-9fe5-b3996d8957d5