Large improvement for the PDB molecule identification code.
This affects the structure.read_pdb user function.  Now the chain ID code, if present in the PDB
file, is being used to determine which ATOM and HETATM records belong to which molecule.  All of the
records for each molecule are stored until the end, when they are all yielded.  This allows for
discontinuous chain IDs throughout the PDB file, something which occurs often with the HETATM
records.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27109 b7916896-f9f9-0310-9fe5-b3996d8957d5