Fixes and improvements for the gnuplot correlation_matrix() plotting API function.
This is for the structure.atomic_fluctuations user function.  The "pm3d map" plot type is incorrect
for such data type, so instead of using 'splot', 'plot' is being used instead.  The resultant EPS
file is now much smaller.  The colour map has also been changed to one of the inbuilt ones for
higher contrast.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27138 b7916896-f9f9-0310-9fe5-b3996d8957d5