Fix for the showing of the spin viewer window in the GUI tests.
The show_tree() method of the main GUI window class was not calling the custom
self.spin_viewer.Show() method, as required to set up the observer objects required to keep the spin
viewer window updated.  The value of status.show_gui was blocking this.  Instead the show argument
of this Show() method is being set to status.show_gui to allow the method to always be executed.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27164 b7916896-f9f9-0310-9fe5-b3996d8957d5