Large speed up of the mol-res-spin selection object.
The Selection.contains_mol(), Selection.contains_res() and Selection.contains_spin() methods of the
lib.selection module have been redesigned for speed.  Instead of setting a number of flags and
performing bit operations at the end of the method to return the correct Boolean value, each of the
multiple checks now simply returns a Boolean value, avoiding all subsequent checks.  The check list
order has also been rearranged so that the least expensive checks are to the top and the most time
intensive checks are last.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27379 b7916896-f9f9-0310-9fe5-b3996d8957d5