Large improvement for the GUI test tearDown() clean up method, fixing the tests on wxPython 2.8.
The user function window destruction has been shifted into a new clean_up_windows() method which is
executed via wx.CallAfter() to avoid racing conditions.  In addition, the spin viewer window is
destroyed between tests.

The spin viewer window change allows the GUI tests to pass on wxPython 2.8 again.  This also allows
the GUI tests to progress much further on Mac OS X systems before they crash again for some other
reason.  This could simply be hiding a problem in the spin viewer window.  However it is likely to
be a racing problem only triggered by the super fast speed of the GUI tests and a normal user would
never be able to operate the GUI on the millisecond timescale and hence may never see it.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27445 b7916896-f9f9-0310-9fe5-b3996d8957d5