Modified the GUI test suite tearDown() method.
The clean_up_windows() method called by tearDown() now prints out a lost of all of the living
windows instead of trying to destroy them (which causes the running of the GUI tests in the GUI to
cause the GUI to be destroyed).  The printouts will be used for debugging purposes.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27455 b7916896-f9f9-0310-9fe5-b3996d8957d5