Fixes for the custom Wiz_window.Destroy() method.
This will now first close the wizard window via the Close() method to make sure all of the wizard
pages are properly updated.  In the end the wizard DestroyChildren() method is called to clean up
all child wx objects, and finally Destroy() is called to eliminate the wizard GUI object.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27457 b7916896-f9f9-0310-9fe5-b3996d8957d5