Improved memory management in the relaxation curve-fitting GUI analysis.
The peak intensity loading wizard is now properly destroyed.  This is both via the delete() function
for terminating the analysis calling the wizard Delete() method, and in the peak_wizard_launch()
method calling the wizard Delete() method prior to overwriting the self.peak_wizard object with a
new GUI wizard.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27467 b7916896-f9f9-0310-9fe5-b3996d8957d5