Huge speed up for the assembly of atomic coordinates from a large number of structures.
The internal structural object validate_models() method was being called once for each structure
when assembling the atomic coordinates.  This resulted in the _translate() internal structural
object method, which converts all input data to formatted strings, being called hundreds of millions
of times.  The problem was in lib.structure.internal.coordinates.assemble_atomic_coordinates(), in
that the one_letter_codes() method, which calls validate_models(), was called for each molecule
encountered.  The solution was not to validate models in one_letter_codes().



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27612 b7916896-f9f9-0310-9fe5-b3996d8957d5