Fix for bug #23293 (https://gna.org/bugs/?23293).
This is the PDB HETATM loading error whereby the last HETATM record is sometimes not read from the
PDB file.  The problem was two-fold.  Firstly the internal structural object _parse_mols_pdb()
method for separating a PDB file into distinct molecules was terminating too early when a new
molecule is found, so that the last PDB record is not appended to the records list for the molecule.
Secondly the write_pdb() method was not handling the PDB sequential serial number correctly.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27673 b7916896-f9f9-0310-9fe5-b3996d8957d5