Created the Structure.test_bug_23294_multi_mol_automerge system test.
This is used to catch bug #2329 (https://gna.org/bugs/?23294), the automatic merging of PDB
molecules resulting in an IndexError.  It reads in the 'in.pdb' PDB file attached to the bug report,
now named 'bug_23294_multi_mol_automerge.pdb', to show the IndexError.  The test also checks the
structure.write_pdb user function to make sure that the output PDB file contains a single merged
molecule.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27680 b7916896-f9f9-0310-9fe5-b3996d8957d5