Fix for bug #23294 (https://gna.org/bugs/?23294).
This is the automatic merging of PDB molecules resulting in an IndexError.  Now if only a single
molecule name is supplied, this will be used for all molecules in the PDB file.  The result is that
the structural data will all be automatically merged into a single molecule.  This merging is
communicated to the user via the current printouts.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27687 b7916896-f9f9-0310-9fe5-b3996d8957d5