Improved the quality of the printouts from the structure.read_pdb user function.
This also affects the structure.read_gaussian and structure.read_xyz user functions.  The messages
about adding new molecules or merging with existing molecules has been significantly improved.  The
text with the model information is now only printed if the model number is present in the PDB file
or has been supplied by the user.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27691 b7916896-f9f9-0310-9fe5-b3996d8957d5