Bug fix for the SHEET PDB records created by the structure.write_pdb user function.
The current and previous atom parts of the record were not being correctly formatted.  This was
simply using the %4s formatting string.  However the PDB atom format is rather more complicated.  To
handle this, the new _handle_atom_name() helper function has been added to the
lib.structure.pdb_write module.  This is now used in the atom() and sheet() functions for
consistently formatting the atom name field.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27700 b7916896-f9f9-0310-9fe5-b3996d8957d5