Added sample data and analysis script, that will eventually show that there is not much differentce in the sample statistics used for comparing the output of two very similar datasets.
This is a multiple comparison test with many t tests at once, where the familywise error is controlled by the Holm method.
Even if the values are very equal, and within the standard deviation, this procedure will reject up to 20% of the null hypothesis.

This is not deemed as a suitable method.

Bug #23644 /https://gna.org/bugs/?23644): monte_carlo.error_analysis() does not update the mean value/expectation value from simulations.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@27852 b7916896-f9f9-0310-9fe5-b3996d8957d5