Cosmetic bug fix for the running of the test suite in the GUI.
The list of skipped tests in the status object was not being reinitialised for each run of the test
suite.  This only affects the GUI where the tests can be run multiple times.  The result was that
the list of skipped tests was always being printed out, even if no tests were skipped.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28010 b7916896-f9f9-0310-9fe5-b3996d8957d5