Manually rotated the frame order matrix element EPS manual figures, for latex2html compatibility.
The '90 rotate' command has been deleted and the bounding box permuted as 'a b c d' -> 'b -c d -a'.
This allows the angle argument in the \includegraphics{} command to be dropped, as latex2html does
not recognise this.  It allows the figures to be visible in the HTML version of the manual at
http://www.nmr-relax.com/manual/index.html .



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28034 b7916896-f9f9-0310-9fe5-b3996d8957d5