The Scons compilation of the PDF and HTML manuals now checks for replicated titles.
A new replicate_title_check target has been added to the scons scripts.  This calls the find()
method of the replicate LaTeX title finding script to determine if any titles are replicated, and if
so the scons target returns with a sys.exit(1) call.  This target is set at the start of the
user_manual_pdf, user_manual_pdf_nofetch, user_manual_html, user_manual_html_nofetch scons targets.

The result is that the manual cannot be compiled if replicate titles exist, forcing the titles to be
changed.  The result will be that the HTML pages at http://www.nmr-relax.com/manual/ will all be
unique, as replicated titles results in only one HTML page being created for all the sections.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28057 b7916896-f9f9-0310-9fe5-b3996d8957d5