Updated all of relax to protect against future changes occurring in the numpy Python package.
From numpy version 1.9, the FutureWarning "__main__:1: FutureWarning: comparison to `None` will
result in an elementwise object comparison in the future." is seen in a large percentage of all
relax's user functions.  This is caught and turned into a RelaxWarning with the same message.

The issue is that the behaviour of the comparison operators '==' and '!=' will change with future
numpy versions.  These have been replaced with 'is' and 'is not' throughout the relax code base.
Changes have also been made to the minfx and bmrblib packages to match.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28063 b7916896-f9f9-0310-9fe5-b3996d8957d5