Modified the internal structural object to be less influenced by the format of the PDB.
The PDB serial number is now intelligently handled, in that it is reset to 1 when a new model is
created.  This information is still kept for supporting the logic of the reading of the CONECT
records, and will be eliminated in the future.

The chain ID information is now no longer stored in the internal structural object, as this
information is recreated by the structure.write_pdb user function based on how the internal
structural object has been created.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28095 b7916896-f9f9-0310-9fe5-b3996d8957d5