Added support for 'observer' structures in the structure.pca user function.
This allows a subset of the structures used in the PC analysis to have zero weight so that these
structures can be used for comparison purposes.  The obs_pipes, obs_models, and obs_molecules
arguments have been added to the user function front end.  The backend uses this to create an array
of weights for each structure.  And the lib.structure.pca functions use the zero weights to remove
the observer structures from the PC mode calculations.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28116 b7916896-f9f9-0310-9fe5-b3996d8957d5