Fix for bug #24601.
This is the failure of the optimisation of the 'R2eff' dispersion model when peaks are missing from
one spectrum, as reported by Petr Padrta at https://gna.org/bugs/?24601.

To handle the missing data, the peak intensity keys are now checked for in the spin container
peak_intensities data structure.  This is both for the R2eff model optimisation as well as the data
back-calculation.  A warning is given when the key is missing.  The relaxation dispersion
base_data_loop() method has been modified to now yield the spin ID string, as this is used in the
warnings.

In addition, the Grace plotting code in the relax library was also modified.  When peak intensity
keys are missing, some of the Grace plots will have no data.  The code will now generate a plot for
that data set, but detect the missing data and allow an empty plot to be created.



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28204 b7916896-f9f9-0310-9fe5-b3996d8957d5