Fix for bug #24723 (https://gna.org/bugs/?24723).
This is the bug that the mean RMSD from the structure.rmsd user function is incorrectly
calculated - it should be a quadratic mean.

The quadratic mean and quadratic standard deviation are now correctly calculated, and the
structure.test_rmsd, structure.test_rmsd_molecules, and structure.test_rmsd_ubi system tests have
been updated for the fix.