Fixes for the Relax_fit.test_inversion_recovery system test.
The wrong equation was used in the calc.py Python script used to calculate the peak intensities in
the test_suite/shared_data/curve_fitting/inversion_recovery/*.list files.  The script and Sparky
files have been updated.  And the I0 value in the script and system test has been changed from 30 to
-30, so that the curves start as negative.

git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28225 b7916896-f9f9-0310-9fe5-b3996d8957d5