Updates to many frame order test suite shared data relax scripts.
These scripts are used for data generation and display, and are not part of the test suite.  The
updates are for the frame_order.pdb_model and pymol.frame_order user functions which no longer
support the 'dist' keyword argument (this functionality was shifted into the frame_order.simulate
user function). 



git-svn-id: svn+ssh://svn.gna.org/svn/relax/trunk@28307 b7916896-f9f9-0310-9fe5-b3996d8957d5